More detailed success-history in timeline
Extend the timeline with a list of dates and links to underline the successes. This is also a good place to link to the pages that have more information about each data sets.

The month for "rnv" is guessed from a date on http://www.gtfs-data-exchange.com/meta/5687256610242560 since I could not find a launch date anywhere else.